rearrange setvar.c
